Extract various subcomponents of /calc and create components/inputs.jsx
* Moved most subcomponents of Calculator in /calc to individual files in
src/subpages/calc/, to clean up /calc.
* Create new components/commonClasses.jsx to house classes commonly
used. For now, it's just those for flex rows and columns, and hidden
elements (using `opacity: 0`).
* Create new components/inputs.jsx to alleviate import hell when dealing
with custom input components. This file currently just re-exports
whatever's inside `components/inputs/`.
