+++ src/components/inputs.jsx
@@ -1,0 +1,5 @@
+import DeckPlanner from './inputs/deckPlanner'
+import NumOfPlayerField from './inputs/numOfPlayerField'
+import SelectField from './inputs/selectField'
+
+export { DeckPlanner, NumOfPlayerField, SelectField }
